Create gateway and interface in publishing module
