+++ src/LearningObjects/Publishing/interactor.ts
@@ -2,6 +2,7 @@ import { LearningObject } from '../../shared/entity';
 import { ResourceError, ResourceErrorReason } from '../../shared/errors';
 import { isAdminOrEditor } from '../../shared/AuthorizationManager';
 import { UserToken } from '../../shared/types';
+import { ReleaseEmailGateway } from './release-email-gateway';
 
 export interface PublishingDataStore {
     addToReleased(releasableObject: LearningObject): Promise<void>;
@@ -12,12 +13,20 @@ export interface PublishingDataStore {
  * then request the data store marks the Learning Object as released. Otherwise, throw a
  * ResourceError.
  */
-export async function releaseLearningObject({ userToken, dataStore, releasableObject }: {
+export async function releaseLearningObject({ userToken, dataStore, releasableObject, releaseEmailGateway }: {
     userToken: UserToken,
     dataStore: PublishingDataStore;
     releasableObject: LearningObject;
+    releaseEmailGateway: ReleaseEmailGateway,
 }): Promise<void> {
     if (isAdminOrEditor(userToken.accessGroups)) {
+        releaseEmailGateway.invokeReleaseNotification({
+            learningObjectName: releasableObject.name,
+            authorName: releasableObject.author.name,
+            collection: releasableObject.collection,
+            authorEmail: releasableObject.author.email,
+            username: releasableObject.author.username,
+        });
         return dataStore.addToReleased(releasableObject);
     }
     throw new ResourceError(`${userToken.username} does not have access to release this Learning Object`, ResourceErrorReason.INVALID_ACCESS);
