+++ src/LearningObjects/Publishing/release-email-gateway.ts
@@ -1,0 +1,9 @@
+export interface ReleaseEmailGateway {
+    invokeReleaseNotification(params: {
+        learningObjectName: string,
+        authorName: string,
+        collection: string,
+        authorEmail: string,
+        username: string,
+    }): void;
+}
